Remove content policy aside
It's a tough call on whether to keep this or not. Ultimately our intuition is saying to remove it for now and then add it back in if we start to see the issue that we're worried about seeing, which is lots of commercial service providers asking to get their service on the page as a means of free advertising.