test: Update `@percy/script` to avoid Percy SDK timeouts (#3804)
When there is a `loading` attribute on an iframe, it causes request interception
to hang in Puppeteer, which caused Percy's asset discovery process to timeout
after 30 seconds (causing missing assets).

This commit updates the SDK to fix the asset discovery timeout (fixed upstream
in Percy's SDKs). It also contains a few small updates to the `snapshots.js`
script to make the tests run faster.