Clarifying scrolling as non-excluded input (#4210)
* Clarifying scrolling as non-excluded input

There's generally confusion around whether or not scrolling is incorporated in CLS calculations. While the spec is clear, this article doesn't mention that scrolling is not excluded input. Hence I've added an aside that mentions it and links to the spec.

Fixes #3835

* updating aside per comments

per philip's comments, updating the hadRecentInput aside to be less generic.

* Update src/site/content/en/metrics/cls/index.md

Co-authored-by: Philip Walton <philipwalton@users.noreply.github.com>

* Update src/site/content/en/metrics/cls/index.md

Co-authored-by: Philip Walton <philipwalton@users.noreply.github.com>

Co-authored-by: Philip Walton <philipwalton@users.noreply.github.com>